Update README and scripts
